Fix Checkstyle warning
