--- domains/tutorials/30-days-of-code/30-interfaces.java
@@ -1,32 +1,0 @@
-import java.util.Scanner;
-
-interface AdvancedArithmetic {
-  int divisorSum(int n);
-}
-//Write your code here
-
-class Calculator implements AdvancedArithmetic {
-  @Override
-  public int divisorSum(int n) {
-    int ret = 0;
-    for (int i = 1; i <= n; i++) {
-      if (n % i == 0) {
-        ret += i;
-      }
-    }
-    return ret;
-  }
-}
-
-class Solution {
-  public static void main(String[] args) {
-    Scanner sc = new Scanner(System.in);
-    int n = sc.nextInt();
-    sc.close();
-
-    AdvancedArithmetic myCalculator = new Calculator();
-    int sum = myCalculator.divisorSum(n);
-    System.out.println("I implemented: " + myCalculator.getClass().getInterfaces()[0].getName());
-    System.out.println(sum);
-  }
-}
